Inicializado según la entrega de enero

+++ src/main/java/org/sunnycake/aton/service/LaboratorioService.java
@@ -1,3 +1,6 @@
+/**
+ * Paquete de lógica del negocio.
+ */
 package org.sunnycake.aton.service;
 
 import java.util.Set;
@@ -6,17 +9,17 @@ import org.sunnycake.aton.dto.Laboratorio;
 
 public interface LaboratorioService {
 
-    Laboratorio buscarLaboratorioPorId(long id);
+    Laboratorio buscarLaboratorioPorId(int id);
 
     void guardarLaboratorio(Laboratorio laboratorio);
 
-    void guardarLaboratorio(long id, String ubicacion, String administracion, String nombre);
+    void guardarLaboratorio(int id, String ubicacion, String administracion, String nombre);
 
     void actualizarLaboratorio(Laboratorio laboratorio);
 
-    void eliminarLaboratorioPorId(long id);
+    void eliminarLaboratorioPorId(int id);
 
     Set<Laboratorio> buscarTodosLosLaboratorios();
 
-    boolean esIdUnico(long id);
+    boolean esIdUnico(int id);
 }
